third_party/WebKit/LayoutTests/css-parser/color-escape-parsing.html

<!doctype html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/property-parsing-test.js"></script>
<script>
assert_valid_value("color", "\\rgb(10, 20, 30)", "rgb(10, 20, 30)");
assert_invalid_value("border-color", "\\#ff0000");
assert_invalid_value("background-color", "rgb\\(255\\,0\\,0\\)");
</script>
